Fixed some dictionaries, did some housekeeping
Moved a lot of the passWORD cracking functions to
"standard_grammar.cpp". This was simply to make things cleaner for
adding new grammar types. Also disabled brute force in password
cracking sessions since that requires a bit of work to move over,
(sloppy coding and global variables), and I want to redo that section
using the OMEN algorithm anyways.