More passphrase work + honeword generation
Mostly messing with the Natural Language Toolkit to generate better
training sets for passphrase creation. Also created a small program to
generate honeywords based on an existing PCFG. The main issue is since
it makes replacements at each stage based on probability, (for example
80% of making a specific replacement), instead of considering the
probabilty as a whole of the final guess, it generates honeywords that
on average a much more "complex" than the training set.