Added basic skeleton for the python pcfg_cracker
I'm starting work on a general purpose python implementation of the
pcfg_cracker. Rather than focus on performance, (I know, the biggest
weakness of the current version), I want to make a very flexible
framework to test out new grammar constructions. Long term I expect the
C++ version to be the applied cracker but the python version to act as
more of a research tool. That assumes I actually get off my butt and
write it though...