Doing more work on the python_pcfg tool. Still not anywhere near done
Finally settled on how to represent the PCFG grammar that I really like.
It's pretty generic so there is a lot of room for new types of rules and
it supports recursion. Right now all the tool does is print out the
terminals associated with a particular pre-terminal. Considering I'm not
even properly loading the grammar, (I'm cheating by reading in a dummy
grammar from a test python file,), it'll be a bit before I'm actually
generating real guesses. But baby steps!