Continued development of python pcfg cracker
Started adding in the priority queue and "next" function so the tool can
start actually generating password guesses from the sample grammar. Very
early on though so there are several major hurdles, (as well as bugs),
that need to be dealt with. This check in is more to show progress and
back up my work than actual features.