The Probability Queue "Next" function is working!
The grammar is working as well as the "next" function so it now acts as
a *VERY* limited password cracker. The biggest limitation is I haven't
added the ability to read in  a realistic PCFG so that's on my to-do
list. Also I'd like to move some of the functionality around to better
differentiate the PCFG and the PQueue so I can make changes to one
without touching the other. That will make future improvements much
easier. So long story short, this is still not something you can use,
but progress is being made ;p