Currently in the middle of adding memory management
I added the first round of memory management for the PQueue. To that
end, I am no longer using the pQueue class and instead using heapq as
that allows me some more flexibility. I currently have it tuned up way
too high as it'll spend most of its time managing memory vs creating
password guesses, but I'm playing around with it and testing it. There
is a lot of optimization that I could focus on but my current plan is
once the memory management is in a stable state I'll move on to the
training program.

Side note, the memory mangement will also make it possible to recover
sessions if they were interrupted as part of it rebuilds the pQueue from
a set point.