Still working on python_pcfg_trainer
Looks like keyboard combos and context sensitive combos are now working.
Added unit tests for both of them. Also did a full rewrite of the
DataList class so that it uses dictionaries as its backend. Finally
calculating the probabilities using the Decimal class since that makes
it easier to enforce the precision that I want for combining items of
the "same" probability.