Most of the basic training is done. Now need to deal with non-ascii characters and save the data
Basic training and probability calculations for Keyboard combos, context
sensitive replacements, digits, letters, capitalization, and special
characters is done. Now I need to save the grammar to disk. During my
unit testing I found that the parsing of non-ASCII characters leaves a
whole lot to be desired, so that is also very high on my to-do list. My
first attempt at using python codecs was mostly a failure as it broke
everything else. Need to research that more.