Training program is fairly stable, now working on the manager
I'm pretty happy where the training program is for now. It's been a long
time since I worked on the pcfg_manager so currently I'm just dusting it
off and standardizing some of the ways I'm doing things between the two
programs. The big next step is to read in the grammar from the training
program