Decided to ditch my previous approach to rebuild the pqueue from max value and use a storage list instead
The downside of this approach is it doesn't support saving/restarting a
session and is still very memory intensive. At a certain point it'll
limit how long a password cracking session can run.  Long story short,
it isn't a full solution, but it'll help for now and it can be later
incorperated into a "build queue from max" solution later. The current
build has settings intentionally low to cause a high degree of thrashing
for testing.