Mostly done with Markov Training
May need to do some work on the top level grammar config file, but for
the most part it looks like the training and inserting of Markov probs
into the main PCFG is working. Now need to focus on the guess generator
(pcfg_manager) to actually create guesses using Markov ranks. May need
to rewrite the guess output portion of it since it currently returns a
list of all guesses for a given pre-terminal as a list. This ... wont
work very well if we have a billion guesses generated by a Markov rank