Markov PoC is working, can make guesses
Downside, current implimentation needs a degree of code cleanup and is
very slow. If left to run for long periods of time, the current guess
framework might not support it as it saves all the guesses for a
pre-terminal in a list before printing them out. That works great if
pre-terminals are averaging around 50 guesses. It works less great if we
get to Markov ranks where millions of guesses are generated.