Making changes to how Markov guesses are generated
Slowly moving the cracker to generate and output guesses vs creating a
full list of all guesses for a pre-terminal and then looping thorugh and
outputing them. Unfortunatly the update for the Markov portion of that
runs real slow so needs to be looked at some more. I'd almost recommend
going back to the previous commit before this if you are cracking
passwords with this right now.