Markov guess generation is working the way I want
Still have code cleanup and would like to get the rest of guess
generation to work the way Markov is, (with a next function vs
generating all the guesses in a list and then printing them out). Also
honeyword generation is currently broken if Markov guesses are included
in the grammar. So there's still some work before I push to Main, but
this commit can crack passwords quite well :)