Lots of work on the new pcfg trainer
Starting to see the light of day with the core functionality of the trainer re-write. Still more to do, and every part I kind of want to spend more time tinkering with, but there's a lot I'm happy with. Also very glad I've been forcing myself to not skip on the unit tests as they have identified a couple of bugs that probably would have gone unnoticed otherwise.