Adding OMEN statistics to better integrate it into a PCFG session
Needs testing, but an eyeball evaluation of the keyspace size and parsing passwords to calculate their OMEN level seems to look like it is working. Adds a few minutes onto the total training session to calculate these for 1 million passwords. "Should" scale well since keyspace calculations are not tied to size of training set and that takes the most relative time. Need to add unit tests and finally integrate OMEN into the pcfg_manager to finalize this.