// WIP: Focusing on the pcfg_guesser
OMEN and normal guess generation is working. Also incorporated the "Deadbeat Dad" algorithm for the priority queue. Moved over the CrackingSession class from version 3.3, and started ripping it up and modifying it for the new pcfg_guesser. Currently it is not generating guesses as I'm playing around with the priority queue code.