Masking an issue when loading an invalid grammar file
Still need to find out why sometimes the tab get's eaten in a "other" grammar file. Suspect a weird character encoding from the training set hides the tab when it is read back in by the guess generator.  Also fixed a bug where if the pcfg_guesser exited but had a pending status output it would print an error to the screen.