Fixes to how input files are parsed for the password scorer
When adding support for count prefixes to the training sets in a previous commit it totally changed how training passwords were read in. I didn't check password scorer, and a gap is I don't have unittests for most of the password scorer code so that totally slipped past me. I updated the code to support the new way to read in passwords.